create webpack config, and file structure
