--- webpack.prod.js
@@ -1,8 +1,0 @@
-const webpack = require("webpack");
-const { merge } = require("webpack-merge");
-const baseConfig = require("./webpack.base");
-
-module.exports = merge(baseConfig, {
-    mode: "production",
-    devtool: "source-map",
-});
