+++ www/styles/pages/index.scss
@@ -1,0 +1,3 @@
+.title {
+    color: green;
+}
